FInal Practica Integradora 3

+++ src/controllers/ProductsController.js
@@ -1,16 +1,16 @@
 import isEmpty from "is-empty";
 import ProductsService from "../services/products.service.js";
-import CommonsUtil from '../utils/Commons.js'
-
+import CommonsUtil from "../utils/Commons.js";
+import { isValidToken } from "../utils/hash.js";
 export default class ProductController {
   static async getProducts(req, res) {
     try {
       const {
         query: { limit = 3, page = 1 },
       } = req;
-      const opts = { limit, page }
-      const products = await ProductsService.paginate( {}, opts);
-      return res.json(CommonsUtil.buidResponse( products ));
+      const opts = { limit, page };
+      const products = await ProductsService.paginate({}, opts);
+      return res.json(CommonsUtil.buidResponse(products));
     } catch (err) {
       return res.status(400).json({
         message: "Error al listar productos",
@@ -46,7 +46,12 @@ export default class ProductController {
   static async addProduct(req, res) {
     try {
       const productData = req.body;
-      await ProductsService.create(productData).catch(() => {
+      const token = await isValidToken(req.cookies.token);
+      const newProductData = {
+        ...productData,
+        owner: token.role === "admin" ? "admin" : token.email,
+      };
+      await ProductsService.create(newProductData).catch(() => {
         throw new Error(
           JSON.stringify({
             detail: "El tipo de dato no es correcto o el código ya existe",
@@ -65,14 +70,23 @@ export default class ProductController {
   static async updateProduct(req, res) {
     try {
       let { pid } = req.params;
+      pid = Number(pid);
       const productData = req.body;
-      await ProductsService.updateOne( pid , productData ).catch(
-        () => {
-          throw new Error(
-            JSON.stringify({ detail: "El tipo de dato no es correcto" })
-          );
-        }
-      );
+      let productById = await ProductsService.getOne(pid);
+      const token = await isValidToken(req.cookies.token);
+      if (token.role === "premium" && token.email !== productById.owner) {
+        throw new Error(
+          JSON.stringify({
+            detail: `El producto fue registrado por otro usuario`,
+          })
+        );
+      }
+
+      await ProductsService.updateOne(pid, productData).catch(() => {
+        throw new Error(
+          JSON.stringify({ detail: "El tipo de dato no es correcto" })
+        );
+      });
       return res.json({
         message: "El producto fue actualizado exitosamente",
       });
@@ -86,30 +100,35 @@ export default class ProductController {
 
   static async deleteProduct(req, res) {
     try {
-      const { pid } = req.params;
-      await ProductsService.deleteById( pid );
+      let { pid } = req.params;
+      pid = Number(pid);
+      const token = await isValidToken(req.cookies.token);
+      let productById = await ProductsService.getOne(pid);
+  
+      if (token.role === "premium" && token.email !== productById.owner) {
+        throw new Error(JSON.stringify( `El producto fue registrado por otro usuario`));
+      }
+      await ProductsService.deleteById(pid);
       return res.json({
         message: "El producto fue eliminado exitosamente",
       });
     } catch (err) {
       return res.status(400).json({
-        message: "Error al buscar el producto",
-        error: JSON.parse(err.message),
+        message: "Error al eliminar el producto",
+        error: err.message,
       });
     }
   }
 
-  static async discountStockProduct(pid , qty) {
-   try {
-    const product = await ProductsService.getById(pid)
-    let stock = product.stock
-    stock -= qty
-    const result = await ProductsService.updateOne( product.id , { stock } )
-    return true
-   } catch (error) {
-    return false
-   }
-
+  static async discountStockProduct(pid, qty) {
+    try {
+      const product = await ProductsService.getById(pid);
+      let stock = product.stock;
+      stock -= qty;
+      const result = await ProductsService.updateOne(product.id, { stock });
+      return true;
+    } catch (error) {
+      return false;
+    }
   }
-
 }
